+++ src/containers/AddMembersToList/selectors.js
@@ -5,14 +5,18 @@
 import { availableEntitiesForList } from '../../redux/modules/entities/selectors';
 import { availablePermissionsForList } from '../../redux/modules/entities/roles/selectors';
 import { entitiesMetaData } from '../../redux/modules/entities/metaData';
+import { availableItemsForList } from '../../redux/modules/entities/listItemSelectors';
 
 export const selectSidePanelTableItems = (state, currentEntity) => {
   const { dependentEntity } = entitiesMetaData[currentEntity];
 
   switch (currentEntity) {
-    case 'outboundIdentifierLists':
-    case 'skills': {
+    case 'outboundIdentifierLists': {
       return availableEntitiesForList(state, dependentEntity, currentEntity);
+    }
+    case 'skills':
+    case 'groups': {
+      return availableItemsForList(state, dependentEntity, currentEntity);
     }
     case 'roles': {
       return availablePermissionsForList(state, dependentEntity, currentEntity);